Create concept of Plugin and have all current plugins (fingerprinters/PBAs) inherit from it. Move plugin manager to their own package. (prep for tests)
