+++ monkey/infection_monkey/post_breach/pba.py
@@ -6,6 +6,7 @@ from infection_monkey.telemetry.post_breach_telem import PostBreachTelem
 from infection_monkey.utils.environment import is_windows_os
 from infection_monkey.config import WormConfiguration
 from infection_monkey.telemetry.attack.t1064_telem import T1064Telem
+from infection_monkey.utils.plugins.plugin import Plugin
 
 LOG = logging.getLogger(__name__)
 
@@ -14,7 +15,7 @@ __author__ = 'VakarisZ'
 EXECUTION_WITHOUT_OUTPUT = "(PBA execution produced no output)"
 
 
-class PBA(object):
+class PBA(Plugin):
     """
     Post breach action object. Can be extended to support more than command execution on target machine.
     """
