+++ monkey/infection_monkey/utils/plugins/plugin.py
@@ -1,0 +1,9 @@
+from abc import ABCMeta, abstractmethod
+
+
+class Plugin(metaclass=ABCMeta):
+
+    @staticmethod
+    @abstractmethod
+    def should_run(class_name: str) -> bool:
+        raise NotImplementedError()
